Updated UI to use 'profile' as home page
