+++ 1-Integrated-Cache/1-2-WebAPI-BgWorker/SPA/src/styles/App.css
@@ -17,7 +17,8 @@ footer {
 }
 
 .navbarButton {
-  color: #fff !important
+  color: #fff !important;
+  padding: .5rem 1rem !important;
 }
 
 .data-area-div {
